+++ cmd/appengine/device.go
@@ -54,7 +54,7 @@ const (
 	-H "User-Agent: astarte-go" \
 	-H "Authorization: Bearer $TOKEN" \
 	"https://$ASTARTE_BASE_URL/appengine/v1/$REALM/devices/$DEVICE_ID/interfaces/$INTERFACE/$PATH
-	-data '{"data" : $DATA}'`
+	--data '{"data" : $DATA}'`
 )
 
 // DevicesCmd represents the devices command
@@ -266,9 +266,7 @@ func printSimpleDevicesList(realm string) {
 		rawPage, _ := deviceListRes.Parse()
 		page, _ := rawPage.([]string)
 
-		for _, deviceID := range page {
-			deviceIDList = append(deviceIDList, deviceID)
-		}
+		deviceIDList = append(deviceIDList, page...)
 	}
 
 	fmt.Println(deviceIDList)
@@ -857,10 +855,8 @@ func devicesGetSamplesF(command *cobra.Command, args []string) error {
 				os.Exit(1)
 			}
 
-			switch rawPage.(type) {
+			switch page := rawPage.(type) {
 			case []client.DatastreamIndividualValue:
-				page, _ := rawPage.([]client.DatastreamIndividualValue)
-
 				// Go with the table header regardless of the requested output type
 				t.AppendHeader(table.Row{"Timestamp", "Value"})
 
@@ -880,8 +876,6 @@ func devicesGetSamplesF(command *cobra.Command, args []string) error {
 				renderOutput(t, sliceAcc, outputType)
 
 			case map[string]client.DatastreamIndividualValue:
-				page, _ := rawPage.(map[string]client.DatastreamIndividualValue)
-
 				// Go with the table header regardless of the requested output type
 				t.AppendHeader(table.Row{"Path", "Timestamp", "Value"})
 
@@ -926,12 +920,10 @@ func devicesGetSamplesF(command *cobra.Command, args []string) error {
 				os.Exit(1)
 			}
 
-			switch rawPage.(type) {
+			switch page := rawPage.(type) {
 			case []client.DatastreamObjectValue:
 				headerRow := table.Row{"Timestamp"}
 				headerPrinted := false
-
-				page, _ := rawPage.([]client.DatastreamObjectValue)
 
 				for _, v := range page {
 					if outputType != "json" {
@@ -968,8 +960,6 @@ func devicesGetSamplesF(command *cobra.Command, args []string) error {
 			case map[string][]client.DatastreamObjectValue:
 				headerRow := table.Row{"Base path", "Timestamp"}
 				headerPrinted := false
-
-				page, _ := rawPage.(map[string][]client.DatastreamObjectValue)
 
 				keys := []string{}
 				for k, v := range page {
@@ -1104,14 +1094,54 @@ func devicesSendDataF(command *cobra.Command, args []string) error {
 		// correctly, we should convert to int every payload for which an integer conversion does not lose
 		// in precision
 		for k, v := range aggrPayload {
+			// since we're dealing with object aggregation, we need to reconstruct
+			// the full path to get the mapping and its type
+			fullPath := fmt.Sprintf("%s/%s", interfacePath, k)
+			mapping, err := interfaces.InterfaceMappingFromPath(iface, fullPath)
+			if err != nil {
+				return err
+			}
+
+			// now we have a mapping type
+			payloadType = mapping.Type
+
 			switch val := v.(type) {
 			case float64:
 				if val == math.Trunc(val) {
 					aggrPayload[k] = int(val)
 				}
+			case string:
+				// in case the type is binaryblob, we want the value as []byte
+				if payloadType == interfaces.BinaryBlob {
+					decoded, err := base64.StdEncoding.DecodeString(string(val))
+					if err != nil {
+						fmt.Fprintln(os.Stderr, "Input string is not base64 encoded.", err)
+						os.Exit(1)
+					}
+					aggrPayload[k] = decoded
+				}
+			case []interface{}:
+				// in case the type is binaryblobarray, we want the values as [][]byte
+				if payloadType == interfaces.BinaryBlobArray {
+					acc := [][]byte{}
+					for _, item := range val {
+						theString, ok := item.(string)
+						if !ok {
+							fmt.Fprintf(os.Stderr, "Invalid item while handling endpoint %s\n", mapping.Endpoint)
+							os.Exit(1)
+						}
+						decoded, err := base64.StdEncoding.DecodeString(theString)
+						if err != nil {
+							fmt.Fprintln(os.Stderr, "Input string is not base64 encoded.")
+							os.Exit(1)
+						}
+
+						acc = append(acc, decoded)
+					}
+					aggrPayload[k] = acc
+				}
 			}
 		}
-
 		parsedPayloadData = aggrPayload
 	}
 
@@ -1294,8 +1324,8 @@ func parseSendDataPayload(payload string, mappingType interfaces.AstarteMappingT
 			return nil, err
 		}
 	case interfaces.BinaryBlob:
-		// We have to verify base64 decoding works
-		if _, err := base64.StdEncoding.DecodeString(payload); err != nil {
+		// if we're dealing with binaryblobs, we want to return a []byte
+		if ret, err = base64.StdEncoding.DecodeString(payload); err != nil {
 			return nil, err
 		}
 	case interfaces.DateTime: