+++ cmd/realm/realm.go
@@ -39,18 +39,18 @@ var astarteAPIClient *client.Client
 func init() {
 	RealmManagementCmd.PersistentFlags().StringP("realm-key", "k", "",
 		"Path to realm private key used to generate JWT for authentication")
-	RealmManagementCmd.MarkPersistentFlagFilename("realm-key")
+	_ = RealmManagementCmd.MarkPersistentFlagFilename("realm-key")
 	RealmManagementCmd.PersistentFlags().String("realm-management-url", "",
 		"Realm Management API base URL. Defaults to <astarte-url>/realmmanagement.")
 	RealmManagementCmd.PersistentFlags().StringP("realm-name", "r", "",
 		"The name of the realm that will be queried")
 	RealmManagementCmd.PersistentFlags().Bool("to-curl", false, "When set, display a command-line equivalent instead of running the command.")
-	viper.BindPFlag("realmmanagement-to-curl", RealmManagementCmd.PersistentFlags().Lookup("to-curl"))
+	_ = viper.BindPFlag("realmmanagement-to-curl", RealmManagementCmd.PersistentFlags().Lookup("to-curl"))
 }
 
 func realmManagementPersistentPreRunE(cmd *cobra.Command, args []string) error {
-	viper.BindPFlag("individual-urls.realm-management", cmd.Flags().Lookup("realm-management-url"))
-	viper.BindPFlag("realm.key-file", cmd.Flags().Lookup("realm-key"))
+	_ = viper.BindPFlag("individual-urls.realm-management", cmd.Flags().Lookup("realm-management-url"))
+	_ = viper.BindPFlag("realm.key-file", cmd.Flags().Lookup("realm-key"))
 	var err error
 	astarteAPIClient, err = utils.APICommandSetup(
 		map[astarteservices.AstarteService]string{astarteservices.RealmManagement: "individual-urls.realm-management"}, "realm.key", "realm.key-file")
@@ -58,7 +58,7 @@ func realmManagementPersistentPreRunE(cmd *cobra.Command, args []string) error {
 		return err
 	}
 
-	viper.BindPFlag("realm.name", cmd.Flags().Lookup("realm-name"))
+	_ = viper.BindPFlag("realm.name", cmd.Flags().Lookup("realm-name"))
 	realm = viper.GetString("realm.name")
 	if realm == "" {
 		return errors.New("realm is required")